:recycle: refactor(DocumentLinkingDialog): Extract PatientPicker.
